重要更新
1、基于iOS7.0开发；
2、移除ASIHTTPRequest；
3、Javascript转Objective-C，支持字符串、数字、布尔、数组、对象、Undefined和Null，共六类数据类型；
4、底层全面升级；

+++ AiitecOpenAPISource/AiitecOpenAPISource/AiitecOpenAPI/AIIPacketRequest/SQLite/Table.m
@@ -56,14 +56,14 @@
 
 - (FMResultSet *)queryResultSet:(NSUInteger)identifier
 {
-    NSString *sql = [NSString stringWithFormat:@"SELECT * FROM %@ WHERE id = %d ", [self tableName], identifier];
+    NSString *sql = [NSString stringWithFormat:@"SELECT * FROM %@ WHERE id = %lu ", [self tableName], identifier];
     return [SQLiteConnection query:sql];
 
 }
 
 - (FMResultSet *)queryResultSet:(NSUInteger)identifier timestampUpdate:(NSString *)timestamp
 {
-    NSString *sql = [NSString stringWithFormat:@"SELECT * FROM %@ WHERE id = %d AND timestamp_update = '%@' ", [self tableName], identifier, timestamp];
+    NSString *sql = [NSString stringWithFormat:@"SELECT * FROM %@ WHERE id = %lu AND timestamp_update = '%@' ", [self tableName], identifier, timestamp];
     return [SQLiteConnection query:sql];
 }
 
@@ -111,13 +111,13 @@
 
 - (int)delete:(NSUInteger)identifier
 {
-    NSString * sql = [NSString stringWithFormat:@"DELETE FROM %@ WHERE id = %d", [self tableName], identifier];
+    NSString * sql = [NSString stringWithFormat:@"DELETE FROM %@ WHERE id = %lu", [self tableName], identifier];
     return [SQLiteConnection update:sql];
 }
 
 - (int)deleteInclude:(AIIModelCollection *)collection
 {
-    int count = [collection count];
+    NSUInteger count = [collection count];
     if (count == 0) {
         return 0;
     }
@@ -126,7 +126,7 @@
     AIIEntity *item = [[AIIEntity alloc] init];
     for (NSUInteger i = 0; i < count; i++) {
         item = [collection objectAtIndex:i];
-        [deleteArray addObject:[NSString stringWithFormat:@"%d", item.identifier]];
+        [deleteArray addObject:[NSString stringWithFormat:@"%lu", item.identifier]];
     }
     NSString *idsDelete = [deleteArray componentsJoinedByString:@","];
     
@@ -137,7 +137,7 @@
 
 - (int)deleteExcept:(AIIModelCollection *)collection
 {
-    int count = [collection count];
+    NSUInteger count = [collection count];
     if (count == 0) {
         return 0;
     }
@@ -146,7 +146,7 @@
     AIIEntity *item = [[AIIEntity alloc] init];
     for (NSUInteger i = 0; i < count; i++) {
         item = [collection objectAtIndex:i];
-        [deleteArray addObject:[NSString stringWithFormat:@"%d", item.identifier]];
+        [deleteArray addObject:[NSString stringWithFormat:@"%lu", item.identifier]];
     }
     NSString *idsDelete = [deleteArray componentsJoinedByString:@","];
     
@@ -156,7 +156,7 @@
 
 - (int)deleteRangeOfIdsExcept:(AIIModelCollection *)collection key:(NSString *)searchKey
 {
-    int count = [collection count];
+    NSUInteger count = [collection count];
     if (count == 0) {
         return 0;
     }
@@ -165,7 +165,7 @@
     AIIEntity *item = [[AIIEntity alloc] init];
     for (NSUInteger i = 0; i < count; i++) {
         item = [collection objectAtIndex:i];
-        [deleteArray addObject:[NSString stringWithFormat:@"%d", item.identifier]];
+        [deleteArray addObject:[NSString stringWithFormat:@"%lu", item.identifier]];
     }
     NSString *idsDelete = [deleteArray componentsJoinedByString:@","];
     
@@ -179,7 +179,7 @@
     NSString *sql = [NSString stringWithFormat:@"DELETE FROM %@ ", [self tableName]];
     NSString *where;
     if (idMin != 0 && idMax !=0) {
-        where = [NSString stringWithFormat:@"WHERE id > %d AND id < %d AND id NOT IN (%@) ", idMin, idMax, idsDelete];
+        where = [NSString stringWithFormat:@"WHERE id > %lu AND id < %lu AND id NOT IN (%@) ", idMin, idMax, idsDelete];
         sql = [sql stringByAppendingString:where];
     }
     if (searchKey.length) {
@@ -241,13 +241,13 @@
 {
     NSMutableArray *mArray = [[NSMutableArray alloc] init];
     AIIEntity *item;
-    int count = [collection count];
+    NSUInteger count = [collection count];
     for (NSUInteger i = 0; i < count; i++) {
         item = [collection objectAtIndex:i];
-        NSString *WHEREString = [NSString stringWithFormat:@"WHERE id = %d AND timestamp_update = '%@'", item.identifier, item.timestampUpdate];
+        NSString *WHEREString = [NSString stringWithFormat:@"WHERE id = %lu AND timestamp_update = '%@'", item.identifier, item.timestampUpdate];
         int n = [SQLiteConnection countWithTableName:[self tableName] where:WHEREString];
         if (n == 0) {
-            [mArray addObject:[NSString stringWithFormat:@"%d", item.identifier]];
+            [mArray addObject:[NSString stringWithFormat:@"%lu", item.identifier]];
         }
     }
     NSString *ids = [mArray componentsJoinedByString:@","];
